separate testing api to its own file
